==> notebook/01_load_and_validate_raw_data.ipynb <==
from pathlib import Path
import requests

import pandas as pd


def download_file(year: int, month: int) -> Path:
    """
    D

    """
    URL = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet"

    resp = requests.get(URL)

    # check to see if get request was sucessful
    if resp.status_code == 200:

        path = f"../data/raw/rides_{year}-{month:02d}.parquet"
        with open(path, 'wb') as f:
            f.write(resp.content)
        
        return path
    else:
        raise Exception(f'{URL} is not available')


download_file(year=2022, month=1)

rides = pd.read_parquet('../data/raw/rides_2022-01.parquet')
rides.head()

columns = ['tpep_pickup_datetime', 'PULocationID' ]
rides = rides[columns]
rides.rename(columns={
    'tpep_pickup_datetime':'pickup_datetime', 
    'PULocationID':'pickup_location_id',
}, inplace=True)
rides.head()

rides['pickup_datetime'].describe()

date_filter_mask = (rides.pickup_datetime >= '2022-01-01') & (rides.pickup_datetime < '2022-02-01')
rides = rides[date_filter_mask]
rides.pickup_datetime.describe()

rides.pickup_location_id.isna().sum()

rides.to_parquet('../data/transformed/validated_rides_2022_01.parquet')
