==> notebook/02_transform_raw_data_into_ts_data.ipynb <==
import pandas as pd

rides  = pd.read_parquet('../data/transformed/validated_rides_2022_01.parquet')
rides.head(10)

rides['pickup_hour'] = rides['pickup_datetime'].dt.floor('h')
rides

agg = rides.groupby(by=['pickup_hour', 'pickup_location_id']).size().reset_index()
agg.rename(columns={0:'rides'}, inplace=True)
agg.head()

agg.columns

from tqdm import tqdm
def add_missing_timeslots(agg_rides: pd.DataFrame) -> pd.DataFrame:

    locations = agg_rides['pickup_location_id'].unique()
    full_range = pd.date_range(
        agg_rides['pickup_hour'].min(), agg_rides['pickup_hour'].max(), freq='h'
    )
    output = pd.DataFrame()
    print(agg_rides.columns)
    for location in tqdm(locations):
    
        agg_rides_i = agg_rides.loc[agg_rides.pickup_location_id == location, ['pickup_hour', 'rides']]

        agg_rides_i.set_index('pickup_hour', inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)

        agg_rides_i['pickup_location_id'] = location

        output = pd.concat([output, agg_rides_i])

    return output.reset_index().rename(columns={'index':'pickup_hour'})

agg_rides_all_slots = add_missing_timeslots(agg)

from typing import Optional, List
import plotly.express as px

def plot_rides(
        rides: pd.DataFrame,
        locations: Optional[List[int]] = None
):
    """
    plot time-series data
    """

    rides_to_plot = rides[rides.pickup_location_id.isin(locations)] if locations else rides

    fig = px.line(
        rides_to_plot,
        x="pickup_hour",
        y='rides',
        color='pickup_location_id',
        template='none',
    )

    fig.show()

plot_rides(agg_rides_all_slots, [43])

agg_rides_all_slots.to_parquet('../data/transformed/ts_data_2022_01.parquet')
