==> notebook/03_transform_timeseries_data_into_features_and_tartgets.ipynb <==
import pandas as pd

ts_data = pd.read_parquet('../data/transformed/ts_data_2022_01.parquet')
# ts_data
ts_one_location = ts_data[ts_data.pickup_location_id == 43]

def get_cutoff_indices(
        data: pd.DataFrame,
        n_features: int,
        step_size: int
) -> list:
    stop_position = len(data) - 1

    subseq_first_idx = 0
    subseq_mid_idx = n_features
    subseq_last_idx = n_features + 1
    indices = []

    while subseq_last_idx <= stop_position:
        indices.append((subseq_first_idx, subseq_mid_idx, subseq_last_idx))

        subseq_first_idx += step_size
        subseq_mid_idx += step_size
        subseq_last_idx += step_size
    
    return indices

n_features = 24
step_size = 1

indices = get_cutoff_indices(
    ts_one_location,
    n_features,
    step_size
)
# indices[:5]

import numpy as np

n_examples = len(indices)
x = np.ndarray(shape=(n_examples, n_features), dtype=np.float32)
y =  np.ndarray(shape=(n_examples), dtype=np.float32)

pickup_hours = []

for i, idx in enumerate(indices):
    x[i, :] = ts_one_location.iloc[idx[0]:idx[1]]['rides'].values
    y[i] = ts_one_location.iloc[idx[1]]['rides']
    pickup_hours.append(ts_one_location.iloc[idx[1]]['pickup_hour'])

features_one_location = pd.DataFrame(
    x,
    columns= [f"rides_previous_{i+1}_hour" for i in reversed(range(n_features))]
)
features_one_location

targets_one_location = pd.DataFrame(y, columns=["target_rides_next-hour"])

from tqdm import tqdm

def transform_ts_data_into_features_and_target(
        ts_data: pd.DataFrame,
        input_seq_len: int,
        step_size: int,
)-> pd.DataFrame:
    """
    Slices and transposes data from time-series format into a (features, target)
    format that we can use to train Supervised ML Models
    """

    assert set(ts_data.columns) == {'pickup_hour', 'rides', 'pickup_location_id'}

    location_ids = ts_data['pickup_location_id'].unique()
    features = pd.DataFrame()
    targets = pd.DataFrame()

    for location_id in tqdm(location_ids):

        ts_data_one_location = ts_data.loc[
            ts_data.pickup_location_id == location_id,
            ['pickup_hour', 'rides']
        ]


        indices = get_cutoff_indices(
            ts_data_one_location,
            input_seq_len,
            step_size
        )

        n_examples = len(indices)
        x = np.ndarray(shape=(n_examples, input_seq_len), dtype=np.float32)
        y =  np.ndarray(shape=(n_examples), dtype=np.float32)

        pickup_hours = []

        for i, idx in enumerate(indices):
            x[i, :] = ts_one_location.iloc[idx[0]:idx[1]]['rides'].values
            y[i] = ts_one_location.iloc[idx[1]]['rides']
            pickup_hours.append(ts_one_location.iloc[idx[1]]['pickup_hour'])

        features_one_location = pd.DataFrame(
            x,
            columns= [f"rides_previous_{i+1}_hour" for i in reversed(range(input_seq_len))]
        )
        features_one_location['pickup_hours'] = pickup_hours
        features_one_location['pickup_location_id'] = location_id

        targets_one_location = pd.DataFrame(y, columns=["target_rides_next_hour"])

        features = pd.concat([features, features_one_location])
        targets = pd.concat([targets, targets_one_location])

    features.reset_index(inplace=True, drop=True)
    targets.reset_index(inplace=True, drop=True)

    return features, targets['target_rides_next_hour']

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*7,
    step_size=24
)

print(f'{features.shape}')
print(f'{targets.shape}')
