==> notebook/05_baseline_model.ipynb <==
import pandas as pd
import os
import sys
script_dir = os.getcwd() # Get the current working directory
sys.path.append(os.path.abspath(os.path.join(script_dir, '..')))

df = pd.read_parquet('../data/transformed/tabular.parquet')
df

from datetime import datetime
from src.data_split import  train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 9, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

import numpy as np

class BaselineModelPreviosHour:
    """
    Prediction = actual demand observed in the last hour
    """
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series ) -> None:
        pass

    def predict( self, X_test: pd.DataFrame)-> np.array:
        return X_test['rides_previous_1_hour']

model = BaselineModelPreviosHour()
predictions = model.predict(X_test)
predictions


from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:4f}')

class BaselineModelPreviousWeek:
    """ 
    Prediction = actual demand observed at t-7 days
    """

    def fit(self, X_train: pd.DataFrame, y_train: pd.Series ) -> None:
        pass

    def predict( self, X_test: pd.DataFrame)-> np.array:
        return X_test[f'rides_previous_{7*24}_hour']

    

model = BaselineModelPreviousWeek()
predictions = model.predict(X_test)

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:4f}')



class BaselineModelLast4Weeks:
    """ 
    Prediction = Average demand observed at t-7 days, t-24 days, t-21 days, and t-28 days
    """

    def fit(self, X_train: pd.DataFrame, y_train: pd.Series ) -> None:
        pass

    def predict( self, X_test: pd.DataFrame)-> np.array:
        return 0.25 * (
            X_test[f'rides_previous_{7*24}_hour'] +
            X_test[f'rides_previous_{2*7*24}_hour'] +
            X_test[f'rides_previous_{3*7*24}_hour'] +
            X_test[f'rides_previous_{4*7*24}_hour']
        )

    

model = BaselineModelLast4Weeks()
predictions = model.predict(X_test)

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:4f}')
