==> notebook/08_light_gbm_model_with_feature_ engineering.ipynb <==
import os
import sys
script_dir = os.getcwd() # Get the current working directory
sys.path.append(os.path.abspath(os.path.join(script_dir, '..')))

import pandas as pd

from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular.parquet')
df

from datetime import datetime
from src.data_split import  train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 9, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

def average_rides_last_4_weeks(df: pd.DataFrame) -> pd.DataFrame:
    """ 
    Averages rides from t-7 days, t-24 days, t-21 days, and t-28 days ago
    """
    df['average_rides_last_4_weeks'] = 0.25 * (
        df[f'rides_previous_{7*24}_hour'] + 
        df[f'rides_previous_{2*7*24}_hour'] + 
        df[f'rides_previous_{3*7*24}_hour'] +
        df[f'rides_previous_{4*7*24}_hour']
    )

    return df


from sklearn.preprocessing import  FunctionTransformer

add_feature_average_rides_last_4_weeks = FunctionTransformer(average_rides_last_4_weeks, validate=False)

add_feature_average_rides_last_4_weeks.fit_transform(X_train)

from sklearn.base import BaseEstimator, TransformerMixin

class TemporalFeatureEngineering(BaseEstimator, TransformerMixin):
    def fit(self, X, y=None):
        return self
    
    def transform(self, X: pd.DataFrame, y=None) -> pd.DataFrame:
        X_ = X.copy()

        # Generate numeric columns from datetime
        X_['hour'] = X_['pickup_hour'].dt.hour
        X_['day_of_week'] = X_['pickup_hour'].dt.day_of_week
        X_['day_of_year'] = X_['pickup_hour'].dt.day_of_week

        return X_.drop(columns=['pickup_hour'])



add_temporal_features = TemporalFeatureEngineering()
add_temporal_features.fit_transform(X_train)

import lightgbm as lgb

from sklearn.pipeline import make_pipeline
pipeline = make_pipeline(
    add_feature_average_rides_last_4_weeks,
    add_temporal_features,
    lgb.LGBMRegressor()
)

pipeline.fit(X_train, y_train)

from sklearn.metrics import mean_absolute_error

predictions = pipeline.predict(X_test)

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:4f}')
