==> notebook/10_backfill_feature_store.ipynb <==
import os
import sys
script_dir = os.getcwd() # Get the current working directory
sys.path.append(os.path.abspath(os.path.join(script_dir, '..')))
from dotenv import load_dotenv
from src.paths import PARENT_DIR
load_dotenv(PARENT_DIR/'.env')

HOPSWORKS_API_KEY = os.environ['HOPSWORKS_API_KEY']
HOPSWORKS_PROJECT_NAME = 'taxidemanpredictor'

from datetime import datetime
import pandas as pd 
from src.data import load_raw_data

# start = 2022
# end = datetime.now().year

# print(f'Downloading raw data from {start} to {end}')

# rides = pd.DataFrame()

# for year in range(start, end + 1):

#     rides_ = load_raw_data(year)

#     rides = pd.concat([rides, rides_])

# rides.shape

# from src.data import transform_raw_data_into_ts_data
# ts_data = transform_raw_data_into_ts_data(rides)

from src.paths import DATA_DIR
# ts_data.to_parquet(DATA_DIR / 'timeseries/ts_data_big.parquet')
ts_data = pd.read_parquet(DATA_DIR / 'timeseries/ts_data_big.parquet')

import hopsworks

project = hopsworks.login(
    project=HOPSWORKS_PROJECT_NAME,
    api_key_value=HOPSWORKS_API_KEY
)

feature_store = project.get_feature_store()

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description="Time-Series data at hourly frequency",
    primary_key = ['pickup_location_id', 'pickup_hour'],
    event_time='pickup_hour',
)

feature_group.insert(ts_data, write_options={'wait_for_job':False})
