==> notebook/11_feature_pipeline.ipynb <==
import os
import sys
script_dir = os.getcwd() # Get the current working directory
sys.path.append(os.path.abspath(os.path.join(script_dir, '..')))

import src.config as config



from datetime import datetime, timedelta

import pandas as pd

current_date = pd.to_datetime(datetime.now().replace(microsecond=0, second=0, minute=0) )
print(f'{current_date=}')

# we fetch raw data for the last 28 days, to add redundancy to our data pipeline
fetch_data_to = current_date
fetch_data_from  =current_date - timedelta(days=28)

from src.data import load_raw_data

def fetch_batch_raw_data(from_date: datetime, to_date:datetime) -> pd.DataFrame:
    """ 
    Simulate production data b y sampling historical data from 52 weeks ago (i.e. 1 year)
    """

    from_date_ = from_date - timedelta(days=7*52)
    to_date_ = to_date - timedelta(days=7*52)
    
    rides = load_raw_data(year=from_date_.year, months=from_date_.month)
    rides = rides[rides.pickup_date >= from_date_]
    rides2 = load_raw_data(year=to_date_.year, months=to_date_.month)
    rides2 = rides2[rides2.pickup_datetime >= to_date_]
    
    rides = pd.concat([rides, rides2])

    rides['pickup_datetime'] += timedelta(days=7*52)

    rides.sort_values(by=['pickup_location_id','pickup_datetime'], inplace=True)
    # rides.drop_duplicates(subset=['pickup_location_id','pickup_datetime'], inplace=True)

    return rides
